requirements.txt
discord.py>=2.4.0
pymongo[srv,aws,ocsp]>=4.9.0
python-dotenv>=1.0.0
